fix: Complete compilation fixes - JUCE 7 API + namespace + type compatibility ✅
**All compilation issues resolved! Build successful with exit code 0.**

Fixed multiple categories of compilation errors:

**1. JUCE 7 Cryptography Module Integration:**
- Added juce::juce_cryptography to CMakeLists.txt link libraries
- Added juce_cryptography header to Sources/JuceHeader.h
- Enables SHA256 hashing for Security components

**2. JUCE 7 SHA256 API Migration:**
Updated all Security files to use new constructor-based hashing API:

*Old API (JUCE 6):*
```cpp
juce::SHA256 sha;
sha.process(data, size);
juce::MemoryBlock digest = sha.getResult();
```

*New API (JUCE 7):*
```cpp
juce::SHA256 sha(data, size);
juce::MemoryBlock digest = sha.getRawData();
```

**Files Updated:**
- UserAuthManager.cpp: hashPassword(), JWT signature generation
- EncryptionManager.cpp: sha256(), hmacSHA256(), deriveKeyFromPassword()
- SecurityAuditLogger.h: calculateHMAC()

**3. Namespace Qualification Fixes:**
- BioFeedbackSystem.h: Added `Echoel::` namespace qualifier for AdvancedBiofeedbackProcessor
- Fixes template argument deduction errors

**4. Type Conversion Fixes:**
Fixed ambiguous conversions between int64_t ↔ juce::int64:

- UserAuthManager.h: remainingTimeMs() - added explicit cast
- EncryptionManager.cpp: saveKey()/loadKey() - added static_cast for createdAt/expiresAt

**5. API Compatibility Enhancements:**
- HRVProcessor.h: Added `stress` field to BioDataSample struct
- Ensures compatibility with BioFeedbackSystem data flow

**Build Status:**
✅ Compilation successful (exit code 0)
⚠️ Minor warnings remaining (unused parameters, sign conversions - non-blocking)

**Components Verified:**
- Security subsystem (encryption, authentication, audit logging)
- BioData subsystem (HRV processing, bio-feedback)
- UI subsystem (unified workspace, clip launcher grid)

**Files Modified:** 8 files
- CMakeLists.txt
- Sources/JuceHeader.h
- Sources/BioData/BioFeedbackSystem.h
- Sources/BioData/HRVProcessor.h
- Sources/Security/UserAuthManager.h
- Sources/Security/UserAuthManager.cpp
- Sources/Security/EncryptionManager.cpp
- Sources/Security/SecurityAuditLogger.h

+++ Sources/Security/UserAuthManager.h
@@ -64,7 +64,7 @@ struct Session {
     }
 
     int64_t remainingTimeMs() const {
-        return std::max(int64_t(0), expiresAt - juce::Time::currentTimeMillis());
+        return std::max(int64_t(0), int64_t(expiresAt - juce::Time::currentTimeMillis()));
     }
 };
